Rename caf::{internal => detail}::connector
