+++ libcaf_net/caf/detail/connector.cpp
@@ -1,0 +1,9 @@
+#include "caf/detail/connector.hpp"
+
+namespace caf::detail {
+
+connector::~connector() {
+  // nop
+}
+
+} // namespace caf::detail
