+++ libcaf_net/caf/detail/connector.hpp
@@ -6,20 +6,21 @@
 
 #include "caf/net/stream_socket.hpp"
 
+#include "caf/detail/net_export.hpp"
 #include "caf/expected.hpp"
 #include "caf/timespan.hpp"
 
 #include <cstdint>
 #include <string>
 
-namespace caf::internal {
+namespace caf::detail {
 
 /// Abstract interface for establishing outbound TCP connections. Inject a
 /// custom implementation via `with_t::client::connector()` to intercept the
 /// connection step (e.g. in unit tests using connected socket pairs).
-class connector {
+class CAF_NET_EXPORT connector {
 public:
-  virtual ~connector() = default;
+  virtual ~connector();
 
   virtual expected<net::stream_socket>
   connect(const std::string& host, uint16_t port, timespan connection_timeout,
@@ -27,4 +28,4 @@ public:
     = 0;
 };
 
-} // namespace caf::internal
+} // namespace caf::detail
